Migrate web-player to typescript
